put both classes in the same file
