+++ pyccl/baryons/__init__.py
@@ -2,4 +2,3 @@ from .baryons_base import *
 from .schneider15 import *
 from .baccoemu_baryons import *
 from .vandaalen19 import *
-from .baccoemu_baryons_dep import *
