+++ pyccl/baryons/baccoemu_baryons.py
@@ -1,7 +1,8 @@
-__all__ = ("BaryonsBaccoemu",)
+__all__ = ("BaryonsBaccoemu", "BaccoemuBaryons")
 
 import numpy as np
 from copy import deepcopy
+from warnings import warn
 
 from .. import Pk2D
 from . import Baryons
@@ -208,3 +209,15 @@ class BaryonsBaccoemu(Baryons):
             raise ValueError(f"Requested scale factor outside the bounds of "
                              f"the emulator: {(a_min, a_max)} outside of "
                              f"{((self.a_min, self.a_max))}")
+
+
+class BaccoemuBaryons(BaryonsBaccoemu):
+    name = 'BaccoemuBaryons'
+
+    def __init__(self, *args, **kwargs):
+        """This throws a deprecation warning on initialization."""
+        warn(f"Class {self.__class__.__name__} will be deprecated. " +
+             f"Please use {BaryonsBaccoemu.__name__} instead.",
+             DeprecationWarning, stacklevel=2)
+        super().__init__(*args, **kwargs)
+    pass
